refactor: modularize QR components, styling
